Script for running pyflakes and pep8
Adding an executable run_tests.py which runs pyflakes and pep8. If we add unit
tests in the future then this is a natural place to run those as well. Static
checks do not presently pass.

This is largely copied from arm's test runner, which in turn was based on
stem's. I'm getting more and more tempted for stem to vend this functionality,
but that's not really appropriate...