Renaming variables and functions to the underscore convention
Python's conventions are to use camelcase for classes (MyClass) but underscore
for everything else (my_function() and my_variable). Actually, just a little
while ago I spent over a week fixing this for arm (I was originally a java
developer, so camelcase was an old habbit). Thankfully exitmap is a *lot*
smaller.

Did bulk renaming with...

  grep -ilr 'oldName' ./* | xargs -i@ sed -i 's/oldName/new_name/g' @