Simplifying exitselector conditionals
It's been too long since my CS courses. This conditionals made my head hurt...

  not ((address and address in desc.address) or (not address))

... is equal to...

  (not (address and address in desc.address)) and address

... which in turn is equal to...

  (not address or address not in desc.address) and address

The initial 'not address' is obviously pointless so this can finally be
distilled to the far more intuitive...

  address and address not in desc.address

The desc.address is a string. There's no reason to do substring comparison here
so swapping to equality...

  address and address != desc.address

I was tempted to just jump to that since it was clearly what the code was going
for, but that initial consitional was so weird I wanted to prove to myself it
was the same. :)