Rewriting get_exits()
This function is essentially doing a series of filters so simply opting for
that instead. The only functional change is that we now only do the
can_exit_to() filter if we have some hosts. From my read the prior code would
reject everything instead which I suspect was a bug.