Add support for a configuration file.
The configuration file makes it possible to store optional command line
options in the file ~/.exitmaprc.  That way, it is no longer necessary
to type them all the time.  This causes another change: The circuit's
first hop is now an optional argument.  If it is not given, single-hop
circuits are used which are unsupported by most relays.