Add module to check for exits running out of fds.
Some relays did not increase their ulimits prior to starting the tor
process which means that they are running low on or out of file
descriptors.  This module attempts to fetch a simple web page and raises
a warning if the page cannot be fetched.  Note that a warning does not
imply that a relay is running out of file descriptors, so you should
manually check the relay before you conclude anything.