Pass torifying execution environments to modules.
So far, a module's network traffic was transparently routed over Tor by
monkey-patching socket.socket.  This was not flexible and is tricky to
maintain.

This commit passes two execution environments (one for Python functions
and one for system commands) to modules.  The environments can be used
as the modules see fit.  It gives modules more flexibility and
simplifies the code.