Add module to test DNSSEC.
The idea is to resolve a domain that has a deliberately broken DNSSEC
record, and see if that works.  If we can resolve it, the exit relay's
resolver does not validate DNSSEC, which is bad.  We don't control
www.dnssec-failed.org, and have to hope that it does what it claims.