Serialize calls to check_finished().
eventhandler.EventHandler.check_finished is called by both the queue
thread and the main thread, but the tear-down operations that it performs
must only happen once.  The call to sys.exit() DOES NOT prevent the
tear-down from happening twice.  (Is the "queue_thread.daemon = False"
line in EventHandler.__init__ perhaps meant to be "= True"?  It wouldn't
fix the race here, but it might prevent other problems, e.g. the program
hanging forever if a module throws an uncaught exception.)  Anyway,
add an explicit lock.