Support nonblocking connections in torsocks.
This was a major revision to the torsocks module, but the _effect_
is simple to explain: a probe module can now do

    s = socket.socket(AF_UNIX, SOCK_STREAM)
    s.setblocking(False)
    err = s.connect_ex(addr)
    if err == errno.EINPROGRESS:
        select.select([s], [], [])

and the select() will return when the connection is complete.  (In a real
module that needs this, one would have _several_ pending connections and
process them as they become available.)  Note that there is a visible
difference from how a normal socket behaves, here: you select for the
socket to become _readable_ before continuing.  A normal socket becomes
_writable_ when the TCP handshake completes.

There's a bunch of secondary work in here in aid of making the emulation
of non-blocking half-open socket behavior as accurate as possible, and
an ancillary change (too difficult to disentangle into its own commit)
in which we make the monkey-patching of socket.socket more robust.  It
is no longer necessary to put back the original socket.socket around
event queue operations, and the monkey-patching will be 100% cleaned
up if a module throws an uncaught exception.