Tell probe() what destinations it can connect to via each exit.
Formerly, only exits that could connect to _all_ of a module's requested
destinations were included in the set of exits to be probed.  Now,
instead, every exit that can connect to _at least one_ of a module's
requested destinations will be included, and probe() receives a new
keyword argument, destinations=, that tells it which destinations it
can use for each exit.  (The destinations list will be a proper subset
of module.destinations, with all hostnames resolved to IP addresses.)

This does not affect the behavior of any existing module, because all
of the existing modules use only a single destination.