Add command line option for random build delays.
This command line option hopefully makes it a little bit more difficult
for vigilant adversaries to discover our scanning activity.  It adds a
random value to the time delay between two subsequent circuit creations.
As a result, if an adversary monitors circuit creation timings, our
signal should be less periodic and thus less obvious.

We implement this feature by exposing a command line argument that
determines a random time delay in the interval [0, DELAY_NOISE).  This
delay is then either added to or subtracted from (both with probability
0.5) the given build delay.  By default, the argument is 0, and
therefore not in effect.